Fix calculating correct DN for nested group search filter (#20)
This fixes a bug I discovered just now, in where the Group DN for the nested search filter is not properly calculate in case the group is in a subdirectory. In that case, search filters using the dn of the group does not work properly.

For example, assume the simplified structure:

ou=Group
cn=Group1,ou=Group
ou=Subgroup,ou=Group
cn=Group2,ou=Subgroup,ou=Group

where Group2 is a member of Group1.

Because of poorly created DN, we assumed that the DN of Group2 is cn=Group2,ou=Group and not ou=Group,ou=Subgroup,cn=Group2. This is PR fixes this problem.

Co-authored-by: Ira Fesefeldt <fesefeldt@cs.rwth-aachen.de>