Add 'exclude groups located outside group unit' (#33)
Make resolving of groups more configurable by adding an option to exclude groups which are located outside of the configured group unit. By default the ldap plugin will return all groups which are found in the group attribute of the user (e.g.: memberOf attribute), this includes groups which are outside of the groups unit (e.g.: ou=Groups). The new option will ensure that only groups are returned which are located under the group unit.

We've decided to create an option for the behaviour to not break existing installations. The option is only available if the custom profile is selected.

Fixes #22 and #14