Cache invalid credentials (#50)
The LDAP plugin caches successful logins, so that using
valid credentials, concrete requests against the LDAP server
are reduced. This is especially important for Subversion,
because a simple commit consists of multiple requests that
should not all be checked using the LDAP server.
But: this only works for successful requests out of the box.
With invalid credentials, a new request is made for every
single request. While this is not that problematic with manual
requests with a mistyped password, this might lead to an
amount of requests when the user has created an api key.
Doing so, the SCM-Manager will still check all realms,
including LDAP. So, a simple SVN commit with an API key can
lead to a not negligible amount of requests to the LDAP.
To avoid this, we introduce a further cache, that keeps
credentials rejected by the LDAP server for some time
(currently 5 seconds) and does not bother the server again,
while this one is cached. The (unknown) password is kept as
a sha256 hash in this cache, to avoid it being flushed to
disk with another cache configuration.